Create db.models.techie_of_the_month
